# Refactor trap