This refactors the `Trap` and `TrapKind`.

- Renames `TrapKind` into `TrapCode`.
- Move `TrapCode::Host` into `Trap`. `Trap` itself is now an `enum` with `Trap::Code(TrapCode)` and `Trap::Host(_)` variants.
- Use new trap implementation in the rest of the codebase.
- Remove old trap implementation.

~~:warning: This is a draft PR because some performance issues must be addressed first. :warning:~~
**edit:** Performance issues resolved in https://github.com/paritytech/wasmi/pull/318/commits/c762a10d82fe5dcc6f25c55a8a730c629a57397b.

## Benchmarks

## Before

```
test bench_tiny_keccak             ... bench:   1,284,815 ns/iter (+/- 67,223)
test bench_tiny_keccak_v1          ... bench:   1,009,335 ns/iter (+/- 61,929)
test count_until                   ... bench:   1,934,947 ns/iter (+/- 145,110)
test count_until_v1                ... bench:   1,749,203 ns/iter (+/- 151,220)
test fac_opt                       ... bench:      24,012 ns/iter (+/- 2,413)
test fac_opt_v1                    ... bench:         736 ns/iter (+/- 95)
test fac_recursive                 ... bench:      25,579 ns/iter (+/- 1,615)
test fac_recursive_v1              ... bench:       1,469 ns/iter (+/- 260)
test host_calls                    ... bench:      78,543 ns/iter (+/- 5,007)
test host_calls_v1                 ... bench:      59,984 ns/iter (+/- 2,640)
test recursive_ok                  ... bench:     534,615 ns/iter (+/- 132,161)
test recursive_ok_v1               ... bench:     394,317 ns/iter (+/- 33,585)
```

## After

```
test bench_tiny_keccak             ... bench:   1,216,187 ns/iter (+/- 70,402)
test bench_tiny_keccak_v1          ... bench:     989,288 ns/iter (+/- 52,336)
test count_until                   ... bench:   1,786,341 ns/iter (+/- 69,680)
test count_until_v1                ... bench:   1,730,468 ns/iter (+/- 67,684)
test fac_opt                       ... bench:      24,379 ns/iter (+/- 4,005)
test fac_opt_v1                    ... bench:         724 ns/iter (+/- 105)
test fac_recursive                 ... bench:      26,380 ns/iter (+/- 4,908)
test fac_recursive_v1              ... bench:       1,437 ns/iter (+/- 201)
test host_calls                    ... bench:      75,301 ns/iter (+/- 10,619)
test host_calls_v1                 ... bench:      58,766 ns/iter (+/- 7,546)
test recursive_ok                  ... bench:     488,383 ns/iter (+/- 80,187)
test recursive_ok_v1               ... bench:     386,891 ns/iter (+/- 24,189)
```

## Summary

Performance is stable for `v1`.
`v0` improved performance for `recursive_ok` from `534,615 ns/iter` down to `488,383 ns/iter`. (~8%)